requirements.txt
pandas>=2.0
numpy>=1.24
PyYAML>=6.0
scikit-learn>=1.3
tqdm>=4.66
torch>=2.2
torch-geometric>=2.5
